refactoring and commenting
