+++ src/main/java/ca/polymtl/metafy/streamerapi/authentication/ApiKeyNotFoundException.java
@@ -1,4 +1,4 @@
-package ca.polymtl.metafy;
+package ca.polymtl.metafy.streamerapi.authentication;
 
 public class ApiKeyNotFoundException extends Exception {
